Add files via upload
This is the current working code that is able to read the data from a digital caliper. It has a smart LED that can identify and select target diameter by utilising a push-button.

Selecting the 1.75mm diameter as a target will allow for a margin of error of 0.05mm, therefore the LED will turn SOLID if the diameter of the filament is between 1.70 - 1.80mm, it also applies with the diameter target of 2.50mm.

The data can also be zero'd to allow flexibility when to start measuring.